Minimized amount JNIEnv attach
Interaction with JNI must attach the current thread (and then detach). Java objects implementation of Drop would also indirectly need to attach the current thread to JNI.

Now, all the interaction with JNI is done on the Credential impl, the needed Java objects are fetched dropped within each function, this ensures that the whole golden path¹ may be executed with a single attach.

¹Create credential, fetch secrets.